feat: Enhance audio synthesis and gameplay mechanics
- Introduced new BPM limits and music configuration structures in audio_synth.h.
- Added support for different music states and scale modes.
- Enhanced telemetry tracking for control and SFX command drops.
- Updated environment generation functions to use double precision for virtualPlayerZ.
- Replaced shader loading with in-memory shader sources for better performance.
- Implemented new gameplay mechanics including chain links for special attacks (beam and bomb).
- Added validation tests for new gameplay features.
- Updated gameplay update functions to accommodate new mechanics and improve performance.

+++ src/environment.h
@@ -89,8 +89,8 @@ typedef struct {
 // Lifecycle & Update API
 void InitEnvironment(EnvironmentSystem *env, uint32_t runSeed);
 void ReseedEnvironment(EnvironmentSystem *env, uint32_t runSeed);
-void UpdateEnvironment(EnvironmentSystem *env, float virtualPlayerZ, float time, float intensity);
-void DrawEnvironment(const EnvironmentSystem *env, Camera3D camera, float virtualPlayerZ);
+void UpdateEnvironment(EnvironmentSystem *env, double virtualPlayerZ, float time, float intensity);
+void DrawEnvironment(const EnvironmentSystem *env, Camera3D camera, double virtualPlayerZ);
 void UnloadEnvironment(EnvironmentSystem *env);
 bool ValidateEnvironmentGenerator(EnvironmentValidationReport *report);
